// types.ts
export interface Class {
  Class_ID: string;
  Year_ID: string;
  Grade_ID: string;
  Class_Name: string;
  Class_Teacher_ID: string;
  Academic_Year_ID: string;
}